Add characters to docs/storydrafts/02-yuko.md
NotablyPrettyCrazy (who doesn't have a GitHub account, thus I'm
uploading it on his behalf) submitted an entire cast of characters to
the currently lacking "Yuko" canon.

As this means a much larger contribution of his compared to the previous
"Delfid" draft, he'll be designated as a co-authod on "Delfid" as well
as a full, second authod for "Yuko".