Document graphics offsets in METAINFO
METAINFO is a format used by the DoomTools autobuild toolkit. Although
many of the tools lack engine-specific support, METAINFO can be used to
keep track of (and hopefully autobuild) Doom engine offsets relating to
the game graphics.

It also distinguishes between which graphics need to be converted to
Doom GFX and which to Doom Flat format.