Initial obsidian documentation backup
A new wiki-like documentation implemented as an Obsidian vault.
Obsidian is proprietary, but at least it's all plain-text and conversion
tooling should be easy to come by down the line.
For reference: <https://obsidian.md/>

This "Wiki" is meant to be a "living" design doc, meaning it may change
as development continues. It also features a much more elaborate lore
than what we can likely cram into the final game, e.g. covering ideas
from the Yuko period in order to inform some of the characters and
themes involved.

It is also as of yet incomplete, likely missing some links ideas and
referenced pages.